backend/src/requirements.txt
django 
django-filter
djangorestframework
markdown
Faker
psycopg2
